renamed Mappers to Scanners
