+++ drm-core/src/main/java/com/nitorcreations/scanners/HierarchyScanner.java
@@ -1,14 +1,14 @@
-package com.nitorcreations.mappers;
+package com.nitorcreations.scanners;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import com.nitorcreations.domain.Link;
 
-public class InheritanceMapper extends AbstractMapper {
+public class HierarchyScanner extends AbstractScanner {
     private final List<Link> links = new ArrayList<>();
 
-    public InheritanceMapper(final List<Class<?>> classes) {
+    public HierarchyScanner(final List<Class<?>> classes) {
         super(classes);
         gatherLinks();
     }
